requirements.txt
Flask==3.0.3
flask_httpauth==4.8.0
python-dotenv==1.0.1
